+++ map-state-properties-mixin/documentation/routes/home-element/examples/manual-fixed/index.js
@@ -1,0 +1,3 @@
+export {default as jsCodeManualFixed} from '!raw-loader!./example';
+export {default as htmlCodeManualFixed} from '!raw-loader!./template';
+export {default as cssCodeManualFixed} from '!raw-loader!./index.css';