+++ map-state-properties-mixin/documentation/routes/home-element/examples/manual/index.js
@@ -1,0 +1,3 @@
+export {default as jsCodeManual} from '!raw-loader!./example';
+export {default as htmlCodeManual} from '!raw-loader!./template';
+export {default as cssCodeManual} from '!raw-loader!./index.css';