+++ map-state-properties-mixin/documentation/routes/home-element/examples/manual/template.js
@@ -1,0 +1,7 @@
+(context, html) => html`
+<div>
+  <p>Attached: ${context.attached}</p>
+  <button on-click="${(e) => context.attachDetach(e)}">Attach/detach</button>
+  <button on-click="${() => context.updateColor()}">Update Color</button>
+</div>
+`;