+++ map-state-properties-mixin/documentation/routes/home-element/examples/mixin/index.js
@@ -1,0 +1,3 @@
+export {default as jsCodeMixin} from '!raw-loader!./example';
+export {default as htmlCodeMixin} from '!raw-loader!./template';
+export {default as cssCodeMixin} from '!raw-loader!./index.css';